moved linked projects table to collection component folder
